Aceleracion se toma como double
Se revisó error donde la velocidad era cero al final de cada via, ahora
solo es cero si al final de la via el semaforo está en rojo

+++ Simulacion/src/ejecucion/Json.scala
@@ -23,7 +23,7 @@ object Json{
       
   case class pametrosSemaforos(minTiempoVerde:Int,maxTiempoVerde:Int,tiempoAmarillo:Int)
   
-  case class pametrosAceleracion(minimo:Int,maximo:Int)
+  case class pametrosAceleracion(minimo:Double,maximo:Double)
   
   case class pametrosDistanciaSemaforos(XSemaforoFrenar:Int,XSemaforoAmarilloContinuar:Int)
