+++ Simulacion/src/vehiculos/Camion.scala
@@ -5,7 +5,7 @@ import movimiento.Velocidad
 import org.jfree.util.ShapeUtilities
 
 
-class Camion (var placaCM:String,velocidad:Velocidad,aceleracion:Int) extends Vehiculo(placaCM)(velocidad,aceleracion){
+class Camion (var placaCM:String,velocidad:Velocidad,aceleracion:Double) extends Vehiculo(placaCM)(velocidad,aceleracion){
     
   var placaGenerada:String = "R"
